Continue working on getting the ssh validation going

+++ builder/destroyer.py
@@ -8,13 +8,16 @@ def destroy(args, cloud):
     """Destroy a previously built environment."""
     with open(args.hosts, 'r+b', 0) as fh:
         servers = yaml.load(fh.read())
-        while servers:
-            kind, server = servers.popitem()
-            print("Destroying server %s, please wait..." % server.name)
-            cloud.delete_server(server.name, wait=True)
-            # Rewrite the file...
-            fh.seek(0)
-            fh.truncate()
-            fh.flush()
-            fh.write(utils.prettify_yaml(servers))
-            fh.flush()
+        if not servers:
+            print("Nothing to destroy.")
+        else:
+            while servers:
+                kind, server = servers.popitem()
+                print("Destroying server %s, please wait..." % server.name)
+                cloud.delete_server(server.name, wait=True)
+                # Rewrite the file...
+                fh.seek(0)
+                fh.truncate()
+                fh.flush()
+                fh.write(utils.prettify_yaml(servers))
+                fh.flush()
